Experiment with hardcoding the run indices.
When I tried putting the horizontal, vertical, and diagonal run indices all in one list, it overwhelmed the compiler. Also, the compiler is much faster if we explicitly declare the type of each list as "[[[Int]]]" instead of relying on the compiler to infer the type. The data does add up to a few megabytes and it takes about 30 seconds to compile, but it saves a ton of tricky runtime calculations.

A possible hybrid approach is to only save recalculated data for the squares that are within 4 squares of an edge of the board. Calculating the run indices for the squares that are more in the center of the board is straightforward.