Use oversized board
Tried a new approach where we expand the board array to have a four-square margin on each side. The value of each square in the margin area is .outofbounds. Doing this way vastly simplifies the logic of generating and checking potential winning runs because we never generate an index that is outside the bounds of the "board.squares" array.