Work in progress on switching BoardView from text to graphics
I did a quick-and-dirty, kludgy implementation of BoardView, using a UITextView, just to get the app started. This commit is the start of a better implementation in which the game board is drawn dynamically. There will be a third implementation that uses better quality graphics, but I will probably keep this dynamically drawn implementation for debugging.