This unit test doesn't always pass.
Because availableMoveIndices is a Set, for which order is not consistent, the min-max algorithm doesn't always choose the same next move in the case where there are other possible next moves with the same min-max score. There is nothing wrong with using a Set, so I don't want to change the implementation. The test should change so it always passes.