Renamed "boardDim" constant to "squaresPerDim".
If we had a rectangular board instead of a square, I would have a lot of constants and variables referring that use the terms, "width" and "height", but since we *do* have a square, I have started to use the term, "dim", to mean, "width or height". "boardDim" was a different way of using the term, "dim", that was causing confusion.