Working in progress on move choosers specific to the player's color.
Doing straight min-max algorithm isn't working well in the early part of the game where a mere three-level min-max tree has over 10 million nodes. We need to implement strategies to reduce the number of moves to be considered, and those strategies aren't going to be the same for the black and white player. Eventually, there might be special end game strategies, too.