Simplify MinMaxMoveChooser.
I realized that instead of always trying to maximize black's score and minimize white's score, we can instead maximize the score of whichever player is making the first move. Doing it this way is also more in line with the way the min-max algorithm is presented in Computer Science courses, etc.