Make stones wider.
I did some research and it turns out Go stones are about the same width as the spacing between the board's grid lines, so that the stones touch each other. When I tried making them the same width as "squareDim", it didn't look good to me, so I made the stones slightly smaller so one can see a small gap between them.