Can place stones by tapping.
This isn't how it will really work, though. What we actually want to do is update the Board model, which will then be reflected in BoardView.

Also, I want to see if I can rewrite this code so that the BoardView coordinate system is more aligned with the board. In our abstract Board model, square (0, 0) is in the upper, left corner the y-axis points down. In the sprite scene space, we are setting the Board's anchor point to be (0.5, 0.5) in the parent scene -- can we make (0, 0) in the Board sprite align with (0, 0) in the model? Is there an option to invert the y-axis in the sprite scene?