Better anchorPoint for board image
Now, the origin, where y=0, is the same for scene and board image. We still have the difference that the scene's y-axis points up while the Board model's row numbers increase as you move down the board -- now that origins are better aligned it's easy to simply negate location.y when trying to figure out what row the user is tapping on.